+++ src/lib/hotbar/index.css
@@ -1,0 +1,67 @@
+#hotbar-image {
+	position: absolute;
+	top: 100%;
+	left: 50%;
+	height: calc(256px * 4);
+	width: calc(256px * 4);
+	transform: translate(calc(-182px * 2), calc(-22px * 4));
+	clip-path: inset(0px calc(74px * 4) calc(234px * 4) 0px);
+}
+
+#hotbar-items-wrapper {
+	position: absolute;
+	top: 100%;
+	left: 50%;
+	height: calc(256px * 4);
+	width: calc(256px * 4);
+	transform: translate(calc(-182px * 2), calc(-22px * 4));
+	clip-path: inset(0px calc(74px * 4) calc(234px * 4) 0px);
+}
+
+#hotbar-highlight {
+	position: absolute;
+	top: 100%;
+	left: 50%;
+	height: calc(256px * 4);
+	width: calc(256px * 4);
+	margin-left: calc(20px * 4 * 4); /* EDIT THIS TO CHANGE WHICH SLOT IS SELECTED */
+	transform: translate(calc((-24px * 2) - (20px * 4 * 4) ), calc((-22px * 4) + (-24px * 4) + 4px)); /* first need to translate up to account for clipping, then account for size of image, then 1px to center vertically over the image*/
+	clip-path: inset(calc(22px * 4) calc(232px * 4) calc(210px * 4) 0px);
+}
+
+.hotbar-item {
+	display: inline-block;
+	height: 32px;
+	width: 64px;
+	margin-top: 12px;
+	margin-left: 12px;
+	margin-right: 4.25px;
+	background-image: url('invsprite.png');
+	background-size: 2048px auto;
+	text-align: right;
+	font-size: 32px;
+	vertical-align: top;
+	padding-top: 32px;
+	color: #ffffff;
+	text-shadow: 2px 2px 0px #3f3f3f;
+	font-family: mojangles, minecraft, monospace;
+}
+
+#hotbar-item-name {
+	color: white;
+	position: absolute;
+	top: 100%;
+	left: 50%;
+	transform: translate(-50%, calc(-170px));
+	margin-top: 0px;
+	text-shadow: rgb(63, 63, 63) 2px 2px 0px;
+	font-family: mojangles, minecraft, monospace;
+	font-size: 24px;
+	text-align: center;
+}
+
+.hotbar-item-name-fader {
+	opacity: 0;
+	transition: visibility 0s, opacity 2s linear;
+	transition-delay: 2s;
+}
